changed GasStation object to support new Dao, updated presenting logic, added new insert logic
