+++ app/src/main/java/com/ktu/nearfuel/network/APIInterface.kt
@@ -1,6 +1,7 @@
 package com.ktu.nearfuel.network
 
 import com.ktu.components.objects.NearestGasStation.NearestGasStationsJsonResponse
+import com.ktu.components.objects.jsonResponses.gasStationResponse.GasStationResponse
 import io.reactivex.Observable
 import retrofit2.http.GET
 import retrofit2.http.Query
@@ -13,6 +14,10 @@ interface APIInterface {
         @Query("type") type: String,
         @Query("key") key: String
     ): Observable<NearestGasStationsJsonResponse>
+
+    @GET("stations")
+    fun getAllGasStations(
+    ): Observable<GasStationResponse>
 }
