+++ app/src/main/java/com/ktu/nearfuel/network/APIURL.kt
@@ -1,5 +1,5 @@
 package com.ktu.nearfuel.network
 
 object APIURL {
-    var API_URL:String = "near.wellbranding.com/api/"
+    var API_URL:String = "http://fuel.wellbranding.com/api/"
 }
