+++ app/src/main/java/com/ktu/nearfuel/ui/main/presenter/MapsNewPresenter.kt
@@ -24,7 +24,8 @@ class MapsNewPresenter<V : MainMVPView> @Inject constructor(
     }
 
     override fun getStationsNearLocation(latLng: LatLng) {
-        getAllGasStationsFromDao()
+        startObservingStationsFromDao()
+        getStationsFromAPI(latLng)
        // getStationsFromMapsAPI(latLng)
 
     }
@@ -64,34 +65,21 @@ class MapsNewPresenter<V : MainMVPView> @Inject constructor(
 //                }
 //            ))
 //    }
-    fun getStationsFromAPi(latLng: LatLng){
+    fun getStationsFromAPI(latLng: LatLng){
         val location = latLng.latitude.toString() + "," + latLng.longitude.toString()
-        disposable.add(apiInterface.getNearestGasStations(
-            location,
-            "distance",
-            "gas_station",
-            "AIzaSyCSXmOpvyXwUwosS07ROsqv0FLICbIYTBo"
+        disposable.add(apiInterface.getAllGasStations(
         )
             .subscribeOn(Schedulers.io())
             .observeOn(AndroidSchedulers.mainThread())
             .subscribe({
-                val gasStationsList: ArrayList<GasStation> = arrayListOf()
-//                for (e in it.results) {
-//                    val gasStation = GasStation()
-//                    gasStation.title = e.name
-//                    gasStation.lat = e.geometry.location.lat.toString()
-//                    gasStation.lng = e.geometry.location.lng.toString()
-//                    gasStationsList.add(gasStation)
-//                    Completable.fromAction {
-//                        gasStationDao.insertGasStation(gasStation)
-//                    }.subscribeOn(Schedulers.io())
-//                        .observeOn(AndroidSchedulers.mainThread()).subscribe({
-//                        }, {
-//
-//                        })
-//                }
+                val gasStationResponse  = it
+                Completable.fromAction {
+                    gasStationDao.insertAllGasStations(gasStationResponse.data)
+                }.subscribeOn(Schedulers.io())
+                    .observeOn(AndroidSchedulers.mainThread()).subscribe({
+                    }, {
 
-                Log.d("response", it.toString())
+                    })
             },
                 {
 
@@ -101,7 +89,7 @@ class MapsNewPresenter<V : MainMVPView> @Inject constructor(
             ))
     }
 
-    fun getAllGasStationsFromDao() {
+    fun startObservingStationsFromDao() {
         gasStationDao.getAllGasStations()
             .observeOn(AndroidSchedulers.mainThread())
             .subscribe({
