+++ components/src/main/java/com/ktu/components/objects/jsonResponses/gasStationResponse/Links.kt
@@ -1,0 +1,8 @@
+package com.ktu.components.objects.jsonResponses.gasStationResponse
+
+data class Links(
+    val first: String,
+    val last: String,
+    val next: String,
+    val prev: Any
+)
