+++ components/src/main/java/com/ktu/components/objects/jsonResponses/gasStationResponse/Meta.kt
@@ -1,0 +1,10 @@
+package com.ktu.components.objects.jsonResponses.gasStationResponse
+data class Meta(
+    val current_page: Int,
+    val from: Int,
+    val last_page: Int,
+    val path: String,
+    val per_page: Int,
+    val to: Int,
+    val total: Int
+)
